Updated the structure fof backend
