+++ backend/Interfaces/IAuthService.cs
@@ -1,0 +1,10 @@
+using didaktos.backend.Models.DTOs;
+
+namespace didaktos.backend.Interfaces
+{
+    public interface IAuthService
+    {
+        Task<HttpResponseDto<object>> RegisterAsync(RegisterRequestDto request);
+        Task<HttpResponseDto<object>> LoginAsync(LoginRequestDto request);
+    }
+}
