+++ backend/Interfaces/IUserRepository.cs
@@ -1,0 +1,11 @@
+using didaktos.backend.Models;
+
+namespace didaktos.backend.Interfaces
+{
+    public interface IUserRepository
+    {
+        Task<User?> GetUserByEmailAsync(string email);
+        Task<User?> GetUserByIdAsync(Guid id);
+        Task<User> CreateUserAsync(User user);
+    }
+}
