--- backend/Models/DTOs/AuthDTOs.cs
@@ -1,30 +1,0 @@
-namespace didaktos.backend.Models.DTOs
-{
-    public class RegisterRequestDto
-    {
-        public string Name { get; set; } = string.Empty;
-        public string Email { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-        public string Role { get; set; } = "student";
-    }
-
-    public class LoginRequestDto
-    {
-        public string Email { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-    }
-
-    public class AuthDataDto
-    {
-        public string Token { get; set; } = string.Empty;
-        public UserDto User { get; set; } = null!;
-    }
-
-    public class UserDto
-    {
-        public Guid Id { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public string Email { get; set; } = string.Empty;
-        public string Role { get; set; } = string.Empty;
-    }
-}
