+++ backend/Models/DTOs/Requests/LoginRequestDto.cs
@@ -1,0 +1,8 @@
+namespace didaktos.backend.Models.DTOs
+{
+    public class LoginRequestDto
+    {
+        public string Email { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+    }
+}
