+++ backend/Models/DTOs/Requests/RegisterRequestDto.cs
@@ -1,0 +1,10 @@
+namespace didaktos.backend.Models.DTOs
+{
+    public class RegisterRequestDto
+    {
+        public string Name { get; set; } = string.Empty;
+        public string Email { get; set; } = string.Empty;
+        public string Password { get; set; } = string.Empty;
+        public string Role { get; set; } = "student";
+    }
+}
