+++ backend/Models/DTOs/UserDto.cs
@@ -1,0 +1,10 @@
+namespace didaktos.backend.Models.DTOs
+{
+    public class UserDto
+    {
+        public Guid Id { get; set; }
+        public string Name { get; set; } = string.Empty;
+        public string Email { get; set; } = string.Empty;
+        public string Role { get; set; } = string.Empty;
+    }
+}
