+++ backend/Models/User.cs
@@ -10,5 +10,4 @@ namespace didaktos.backend.Models
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
-
 }
